@@ -1,36 +1,34 @@
-import { Component, OnInit, EventEmitter,Input } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/interval';
-import 'rxjs/add/operator/takeWhile'
+import { Component, OnInit, EventEmitter, Input } from "@angular/core";
+import { HttpClient } from "@angular/common/http";
+import { Observable } from "rxjs/Observable";
+import "rxjs/add/observable/interval";
+import "rxjs/add/operator/takeWhile";
 
 @Component({
-  selector: 'app-video',
-  templateUrl: './video.component.html',
-  styleUrls: ['./video.component.css']
+  selector: "app-video",
+  templateUrl: "./video.component.html",
+  styleUrls: ["./video.component.css"],
 })
 export class VideoComponent implements OnInit {
-  @Input() muted:boolean;
-  mobile:boolean=false
-  myVar:boolean = false;
-  videos:string="http://206.189.79.221/video"
+  @Input() muted: boolean = true;
+  mobile: boolean = false;
+  myVar: boolean = false;
+  videos: string = "";
 
-  constructor(private httpClient:HttpClient) {
-    
-   }
-
+  public innerWidth: any;
 
   ngOnInit() {
+    this.innerWidth = window.innerWidth;
+    if (innerWidth < 499) {
+      this.mobile = true;
+      this.videos = "";
+    } else {
+      this.videos = "https://mariachilosnungaray.com.mx/video";
+    }
     Observable.interval(1000)
-    .takeWhile(() => this.myVar = true)
-    .subscribe(i => { 
+      .takeWhile(() => (this.myVar = true))
+      .subscribe((i) => {
         // This will be called every 10 seconds until `stopCondition` flag is set to true
-    })
-    
+      });
   }
-
- 
-
-
-
 }
